feat: Adding hash command
A common use case is to hash some file or value from the command
line. Stuff like sha1 is readily available, but some methods like
keccak or ripemd160 are a little harder to come by. I'm adding this to
create some hash utilities and also just get used to cobra.